# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "The error schema of docs/spec/part1.md section 7.5.",
        "examples": [
          {
            "error": {
              "code": "input_too_large",
              "detail": {
                "limit_bytes": 26214400
              },
              "docs": "https://ezmd.example.com/docs/errors#input_too_large",
              "message": "Upload exceeds the 25 MB limit for anonymous requests.",
              "request_id": "req_3fQ9xL2mVb7Kp1Zt",
              "status": 413
            }
          }
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "JobOut": {
        "description": "A job's state and metadata (`GET /v1/jobs/{job_id}`).",
        "examples": [
          {
            "converter_id": "text.markdown",
            "created_at": "2026-10-12T18:40:00Z",
            "expires_at": "2026-10-13T18:40:00Z",
            "id": "job_7Kx2mQp9Lw3nRt5vYb8cDe",
            "input": {
              "display": "notes.md",
              "kind": "bytes",
              "mime": "text/markdown",
              "size_bytes": 2048
            },
            "profile": "full",
            "progress": 100,
            "queue": "default",
            "stage_message": "Done",
            "state": "done",
            "truncated": false,
            "updated_at": "2026-10-12T18:40:02Z",
            "warnings": [
              "encoding_uncertain"
            ],
            "warnings_count": 1
          }
        ],
        "properties": {
          "converter_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "documents.docling_pdf"
            ],
            "title": "Converter Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Present in failed: the error object (7.5)",
            "title": "Error"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When the job and its results are purged",
            "title": "Expires At"
          },
          "id": {
            "description": "Unguessable job id; possession is the capability",
            "examples": [
              "job_7Kx2mQp9Lw3nRt5vYb8cDe"
            ],
            "title": "Id",
            "type": "string"
          },
          "input": {
            "$ref": "#/components/schemas/InputOut"
          },
          "needs_action": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Present in needs_user_action: what the client must supply and where",
            "title": "Needs Action"
          },
          "profile": {
            "description": "Requested profile",
            "examples": [
              "rag"
            ],
            "title": "Profile",
            "type": "string"
          },
          "progress": {
            "description": "Percent complete",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress",
            "type": "integer"
          },
          "queue": {
            "description": "default, media, fetch, or fetch_residential",
            "examples": [
              "default"
            ],
            "title": "Queue",
            "type": "string"
          },
          "stage_message": {
            "description": "Human-readable stage",
            "examples": [
              "Converting page 84 of 200"
            ],
            "title": "Stage Message",
            "type": "string"
          },
          "state": {
            "enum": [
              "queued",
              "fetching",
              "converting",
              "rendering",
              "done",
              "failed",
              "needs_user_action",
              "expired"
            ],
            "title": "State",
            "type": "string"
          },
          "truncated": {
            "default": false,
            "description": "True when part of the input or output was cut",
            "title": "Truncated",
            "type": "boolean"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "warnings": {
            "description": "Canonical warning codes the conversion emitted, first-seen order (see GET /v1/warnings)",
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          },
          "warnings_count": {
            "default": 0,
            "description": "Number of warnings the conversion emitted",
            "title": "Warnings Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "state",
          "progress",
          "stage_message",
          "queue",
          "input",
          "profile"
        ],
        "title": "JobOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "description": "Optional API key (`ak_<env>_...`). Keyed callers get the key's limits and skip Turnstile.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "FetchNodeBearer": {
        "description": "EZMD_FETCH_NODE_SECRET; requests must also come from EZMD_FETCH_NODE_CIDR.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "ezmd",
      "url": "https://ezmd.example.com/"
    },
...
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "ezmd API",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/jobs/{job_id}": {
      "delete": {
        "description": "Deletes the job row, its input, IR, rendered results, attachments, and buffered events.",
        "operationId": "deleteJob",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Purged"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`invalid_request`: Malformed body, unknown option, or invalid profile override."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`unauthorized`: Missing or unknown API key, or bad fetch-node secret."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`forbidden`: The key or node may not do this (for example residential fetches)."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`not_found`: Unknown job, attachment, or route."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`internal_error`: Unexpected server error (logged with the request id)."
          }
        },
        "security": [
          {},
          {
            "ApiKey": []
          }
        ],
        "summary": "Purge the job and its blobs now",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "description": "State, progress, input metadata, warning codes, and, depending on the state, `needs_action` (what the client must supply) or `error`. Unknown ids and ids owned by another API key both return 404.",
        "operationId": "getJob",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`invalid_request`: Malformed body, unknown option, or invalid profile override."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`unauthorized`: Missing or unknown API key, or bad fetch-node secret."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`forbidden`: The key or node may not do this (for example residential fetches)."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`not_found`: Unknown job, attachment, or route."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`internal_error`: Unexpected server error (logged with the request id)."
          }
        },
        "security": [
          {},
          {
            "ApiKey": []
          }
        ],
        "summary": "Job state and metadata",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/attachments/{path}": {
      "get": {
        "description": "A file the conversion produced (table CSVs, extracted images), by the relative path listed in the sidecar. Images are served inline; everything else as an attachment with `nosniff`.",
        "operationId": "getJobAttachment",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "The attachment bytes"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`invalid_request`: Malformed body, unknown option, or invalid profile override."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`unauthorized`: Missing or unknown API key, or bad fetch-node secret."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`forbidden`: The key or node may not do this (for example residential fetches)."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`not_found`: Unknown job, attachment, or route."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`rate_limited`: Too many requests, jobs, or concurrent jobs; see Retry-After."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`internal_error`: Unexpected server error (logged with the request id)."
          }
        },
        "security": [
          {},
          {
            "ApiKey": []
          }
        ],
        "summary": "CSV sidecars and extracted images",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/events": {
      "get": {
        "description": "A `text/event-stream` of numbered events: `state`, `progress`, `warning` (a Warning JSON), and one terminal `done`, `failed`, or `needs_user_action`, after which the stream closes. A `: keepalive` comment is sent every 15 s. Reconnect with `Last-Event-ID` (or `last_event_id`) to resume; the buffer keeps 500 events.",
        "operationId": "streamJobEvents",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "description": "Alternative to the Last-Event-ID header",
            "in": "query",
            "name": "last_event_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Alternative to the Last-Event-ID header",
              "title": "Last Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "example": "event: progress\nid: 4\ndata: {\"progress\":42,\"stage_message\":\"Converting page 84 of 200\"}\n",
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "The event stream"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`invalid_request`: Malformed body, unknown option, or invalid profile override."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`unauthorized`: Missing or unknown API key, or bad fetch-node secret."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`forbidden`: The key or node may not do this (for example residential fetches)."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`not_found`: Unknown job, attachment, or route."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`rate_limited`: Too many requests, jobs, or concurrent jobs; see Retry-After."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`internal_error`: Unexpected server error (logged with the request id)."
          }
        },
        "security": [
          {},
          {
            "ApiKey": []
          }
        ],
        "summary": "Server-sent events: progress, state, warning, done, failed, needs_user_action",
        "tags": [
          "jobs"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "This instance",
      "url": "https://ezmd.example.com"
    }
  ],
...
}